refactor: move common code out
